check if the db exist

+++ src/lib.rs
@@ -1,9 +1,20 @@
 use ini::Ini;
 use std::path::Path;
 
+fn exist(path:&str)->bool{
+    let created:bool=false;
+    if Path::new(path).exists() == true{
+        created=false;
+    }else{
+        created=true;
+    }
+    
+    created
+}
+
 pub fn new(path:&str){
 
-    if Path::new(path).exists() == true{
+    if exist(path) == true{
         println!("the db already exist");
     }else{
         let db=Ini::new();
@@ -12,20 +23,24 @@ pub fn new(path:&str){
 }
 
 pub fn insert(path:&str, data:Vec<(&str, &str)>){
+    if exist(path)==false{
+        println!("the db don't exist yet !!!");
+    }else{
+        let mut db=Ini::load_from_file(path).unwrap();
 
-    let mut db=Ini::load_from_file(path).unwrap();
 
+        let mut id=db.len();
 
-    let mut id=db.len();
+        for object in data.iter(){
+            db.with_section(Some((id).to_string()))
+                .set(object.0.to_string(), object.1.to_string());     
+        }
 
-    for object in data.iter(){
-        db.with_section(Some((id).to_string()))
-            .set(object.0.to_string(), object.1.to_string());     
+        db.write_to_file(path).unwrap();
+
+        id+=1;
     }
-
-    db.write_to_file(path).unwrap();
-
-    id+=1;
+    
 }
 
 pub fn get_db(path:&str)->Ini{
@@ -35,40 +50,53 @@ pub fn get_db(path:&str)->Ini{
 
 pub fn update(path:&str, data:Vec<(&str, &str)>,pos:i32){
 
-    let mut db=Ini::load_from_file(path).unwrap(); 
+    if exist(path)==false{
+        println!("the db don't exist yet !!!");
+    }else{
 
-    for object in data.iter(){
+        let mut db=Ini::load_from_file(path).unwrap(); 
 
-        let id=pos;
+        for object in data.iter(){
 
-        db.with_section(Some((id).to_string()))
-            .set(object.0.to_string(), object.1.to_string());
+            let id=pos;
+
+            db.with_section(Some((id).to_string()))
+                .set(object.0.to_string(), object.1.to_string());
+        }
+
+        db.write_to_file(path).unwrap();
+
     }
-
-    db.write_to_file(path).unwrap();
 }
 
 pub fn delete(path:&str, pos:i32){
 
-    let mut db=Ini::load_from_file(path).unwrap(); 
+    if exist(path)==false{
+        println!("the db don't exist yet !!!")
+    }else{
+        let mut db=Ini::load_from_file(path).unwrap(); 
 
-    db.delete(Some((pos).to_string()));
+        db.delete(Some((pos).to_string()));
 
-    db.write_to_file(path).unwrap();
+        db.write_to_file(path).unwrap();
 
+    }
 }
 
 pub fn delete_all(path:&str){
+    if exist(path)==false{
+        println!("the db don't exist yet !!!")
+    }else{
+        let mut db=Ini::load_from_file(path).unwrap(); 
+        
+        for i in 0..db.len(){
+            db.delete(Some((i).to_string()));
+        }
 
-    let mut db=Ini::load_from_file(path).unwrap(); 
-    
-    for i in 0..db.len(){
-        db.delete(Some((i).to_string()));
+
+        db.write_to_file(path).unwrap();
+
     }
-
-
-    db.write_to_file(path).unwrap();
-
 }
